+++ FESBCompanion/Sources/Core/Encodable+Extensions.swift
