+++ FESBCompanion/Sources/Core/Int+Extensions.swift
