+++ FESBCompanion/Sources/Core/String+Extensions.swift
