+++ FESBCompanion/Sources/Core/Weekday.swift
@@ -1,0 +1,5 @@
+public enum Weekday: Int {
+
+    case sunday = 1, monday, tuesday, wednesday, thursday, friday, saturday
+
+}
